+++ Libraries/LibWeb/CSS/CSSPageRule.h
@@ -8,31 +8,9 @@
 
 #include <LibWeb/CSS/CSSGroupingRule.h>
 #include <LibWeb/CSS/CSSPageDescriptors.h>
+#include <LibWeb/CSS/PageSelector.h>
 
 namespace Web::CSS {
-
-enum class PagePseudoClass : u8 {
-    Left,
-    Right,
-    First,
-    Blank,
-};
-Optional<PagePseudoClass> page_pseudo_class_from_string(StringView);
-StringView to_string(PagePseudoClass);
-
-class PageSelector {
-public:
-    PageSelector(Optional<FlyString> name, Vector<PagePseudoClass>);
-
-    Optional<FlyString> name() const { return m_name; }
-    Vector<PagePseudoClass> const& pseudo_classes() const { return m_pseudo_classes; }
-    String serialize() const;
-
-private:
-    Optional<FlyString> m_name;
-    Vector<PagePseudoClass> m_pseudo_classes;
-};
-using PageSelectorList = Vector<PageSelector>;
 
 // https://drafts.csswg.org/css-page-3/#at-ruledef-page
 class CSSPageRule final : public CSSGroupingRule {
